requirements.txt
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.4
django-crispy-forms==1.9.1
django-mathfilters==1.0.0
djangorestframework==3.11.0
django-filter==2.3.0
djangorestframework-jwt==1.11.0
django-cors-headers==3.3.0
six==1.15.0
gunicorn==20.0.4
idna==2.9
public==2019.4.13
pytz==2020.1
query-string==2019.4.13
requests==2.24.0
urllib3==1.25.9
whitenoise==5.1.0
django-heroku
PyJWT==1.7.1
django-rest-registration==0.5.6